start breaking up ExpDict into small files
